Complete instructions for TOP Recipes project
Followed the instructions for The Odin Project Recipes project.
This included creating index.html with links to three recipe pages.
Each recipe page includes an image, the ingredients list, and the steps
required to cook each recipe.

The intent of this project is to practice creating pages with links
and images.